+++ aderyn_core/src/ast/impls/node/mod.rs
@@ -1,1 +1,2 @@
 pub mod blocks;
+pub mod contracts;
