--- aderyn_core/src/ast/raw_nodes.rs
@@ -1,8 +1,0 @@
-use super::{macros::ast_node, Statement};
-use serde::{Deserialize, Serialize};
-
-ast_node!(
-    struct Block {
-        statements: Vec<Statement>,
-    }
-);
